Run test app before running playwright tests
