+++ playwright.config.js
@@ -6,7 +6,7 @@ const { devices } = require('@playwright/test');
  * @type {import('@playwright/test').PlaywrightTestConfig}
  */
 const config = {
-  testDir: './playwright',
+  testDir: './test/playwright',
   /* Maximum time one test can run for. */
   timeout: 30 * 1000,
   expect: {
@@ -28,10 +28,10 @@ const config = {
   reporter: 'html',
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
+    /* Base URL to use in actions like `await page.goto('/')`. */
+    baseURL: 'http://localhost:1337',
     /* Maximum time each action such as `click()` can take. Defaults to 0 (no limit). */
     actionTimeout: 0,
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    // baseURL: 'http://localhost:3000',
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
@@ -93,10 +93,12 @@ const config = {
   outputDir: 'test-results/',
 
   /* Run your local dev server before starting the tests */
-  // webServer: {
-  //   command: 'npm run start',
-  //   port: 3000,
-  // },
+  webServer: {
+    command: 'yarn test:generate-app test-apps/playwright --run',
+    port: 1337,
+    timeout: 30 * 1000,
+    reuseExistingServer: !process.env.CI,
+  },
 };
 
 module.exports = config;
